// src/types/weather.ts
export interface WeatherInfo {
  cityName: string;
  country: string;
  weatherDescription: string;
  icon: string;
  day: string;
  time: string;
  sunrise: string;
  sunset: string;
  visibility: string;
  pressure: string;
  temp: string;
  feelsLike: string;
  tempMin: string;
  tempMax: string;
  humidity: string;
  weather: string;
  windSpeed: string;
}